refactor(machine): unify server and trail features under machine domain
- Merge features/server into features/machine (controller, service,
  storage, types)
- Merge features/trail into features/machine as trial sub-feature
- Consolidate all routes under /api/v1/machines prefix
- Update frontend API paths from v1/servers and v1/machine/ to v1/machines/
- Update SSH banner hook to use /v1/machines/ssh/status
