+++ api/internal/features/machine/storage/list.go
@@ -7,48 +7,44 @@ import (
 	"strings"
 
 	"github.com/google/uuid"
-	"github.com/nixopus/nixopus/api/internal/features/server/types"
+	"github.com/nixopus/nixopus/api/internal/features/machine/types"
 	shared_types "github.com/nixopus/nixopus/api/internal/types"
 	"github.com/uptrace/bun"
 )
 
-type ServerStorage struct {
+type ListStorage struct {
 	DB  *bun.DB
 	Ctx context.Context
 	tx  *bun.Tx
 }
 
-func (s *ServerStorage) getDB() bun.IDB {
+func NewListStorage(db *bun.DB, ctx context.Context) *ListStorage {
+	return &ListStorage{DB: db, Ctx: ctx}
+}
+
+func (s *ListStorage) getDB() bun.IDB {
 	if s.tx != nil {
 		return *s.tx
 	}
 	return s.DB
 }
 
-// ServerRepository defines the interface for server storage operations.
-// This enables mocking in tests.
-type ServerRepository interface {
-	ListServersByOrganizationID(orgID uuid.UUID, params types.ServerListParams) ([]types.ServerResponse, int, error)
-	SetDefaultServer(orgID uuid.UUID, serverID uuid.UUID) (*uuid.UUID, error)
-	GetServerByIDAndOrgID(serverID uuid.UUID, orgID uuid.UUID) (*shared_types.SSHKey, error)
+type ListRepository interface {
+	ListMachinesByOrganizationID(orgID uuid.UUID, params types.MachineListParams) ([]types.MachineResponse, int, error)
+	SetDefaultMachine(orgID uuid.UUID, machineID uuid.UUID) (*uuid.UUID, error)
+	GetMachineByIDAndOrgID(machineID uuid.UUID, orgID uuid.UUID) (*shared_types.SSHKey, error)
 }
 
-// ListServersByOrganizationID retrieves all SSH keys (servers) for an organization
-// with optional user_provision_details joined as provision field.
-// Returns servers, total count, and error.
-func (s *ServerStorage) ListServersByOrganizationID(orgID uuid.UUID, params types.ServerListParams) ([]types.ServerResponse, int, error) {
-	// Build base query for SSH keys
+func (s *ListStorage) ListMachinesByOrganizationID(orgID uuid.UUID, params types.MachineListParams) ([]types.MachineResponse, int, error) {
 	query := s.getDB().NewSelect().
 		TableExpr("ssh_keys AS sk").
 		Where("sk.organization_id = ?", orgID).
 		Where("sk.deleted_at IS NULL")
 
-	// Apply is_active filter if provided
 	if params.IsActive != nil {
 		query = query.Where("sk.is_active = ?", *params.IsActive)
 	}
 
-	// Apply search filter
 	if params.Search != "" {
 		searchPattern := "%" + strings.ToLower(params.Search) + "%"
 		query = query.WhereGroup(" AND ", func(q *bun.SelectQuery) *bun.SelectQuery {
@@ -58,7 +54,6 @@ func (s *ServerStorage) ListServersByOrganizationID(orgID uuid.UUID, params type
 		})
 	}
 
-	// If status filter is provided, filter by user.provision_status
 	if params.Status != "" {
 		query = query.Join("INNER JOIN user_provision_details AS upd ON sk.id = upd.ssh_key_id").
 			Join("INNER JOIN \"user\" AS u ON upd.user_id = u.id").
@@ -66,7 +61,6 @@ func (s *ServerStorage) ListServersByOrganizationID(orgID uuid.UUID, params type
 			Where("u.provision_status = ?", params.Status)
 	}
 
-	// Build count query for total count
 	countQuery := s.getDB().NewSelect().
 		TableExpr("ssh_keys AS sk").
 		Where("sk.organization_id = ?", orgID).
@@ -85,7 +79,6 @@ func (s *ServerStorage) ListServersByOrganizationID(orgID uuid.UUID, params type
 		})
 	}
 
-	// If status filter is provided, filter count by user.provision_status
 	if params.Status != "" {
 		countQuery = countQuery.Join("INNER JOIN user_provision_details AS upd ON sk.id = upd.ssh_key_id").
 			Join("INNER JOIN \"user\" AS u ON upd.user_id = u.id").
@@ -93,16 +86,13 @@ func (s *ServerStorage) ListServersByOrganizationID(orgID uuid.UUID, params type
 			Where("u.provision_status = ?", params.Status)
 	}
 
-	// Get total count
 	totalCount, err := countQuery.Count(s.Ctx)
 	if err != nil {
 		return nil, 0, err
 	}
 
-	// Apply sorting
 	sortColumn := "sk.created_at"
 	if params.SortBy != "" {
-		// Validate sort column to prevent SQL injection
 		validSortColumns := map[string]string{
 			"name":       "sk.name",
 			"created_at": "sk.created_at",
@@ -121,43 +111,32 @@ func (s *ServerStorage) ListServersByOrganizationID(orgID uuid.UUID, params type
 
 	query = query.OrderExpr("? ?", bun.Ident(sortColumn), bun.Safe(sortOrder))
 
-	// Apply pagination
 	offset := (params.Page - 1) * params.PageSize
-	query = query.Limit(params.PageSize).Offset(offset)
+	query = query.Limit(params.PageSize).Offset(offset).ColumnExpr("sk.*")
 
-	// Select SSH key columns
-	query = query.ColumnExpr("sk.*")
-
-	// Execute query to get SSH keys
 	var sshKeys []shared_types.SSHKey
-	err = query.Scan(s.Ctx, &sshKeys)
-	if err != nil {
+	if err = query.Scan(s.Ctx, &sshKeys); err != nil {
 		return nil, 0, err
 	}
 
 	if len(sshKeys) == 0 {
-		return []types.ServerResponse{}, totalCount, nil
+		return []types.MachineResponse{}, totalCount, nil
 	}
 
-	// Get SSH key IDs
 	sshKeyIDs := make([]uuid.UUID, 0, len(sshKeys))
 	for _, key := range sshKeys {
 		sshKeyIDs = append(sshKeyIDs, key.ID)
 	}
 
-	// Query user_provision_details for these SSH keys
 	var provisionDetails []shared_types.UserProvisionDetails
-	provisionQuery := s.getDB().NewSelect().
+	if err = s.getDB().NewSelect().
 		Model(&provisionDetails).
 		Where("ssh_key_id IN (?)", bun.In(sshKeyIDs)).
-		Where("organization_id = ?", orgID)
-
-	err = provisionQuery.Scan(s.Ctx)
-	if err != nil {
+		Where("organization_id = ?", orgID).
+		Scan(s.Ctx); err != nil {
 		return nil, 0, err
 	}
 
-	// Create a map of provision details by SSH key ID
 	provisionMap := make(map[uuid.UUID]*shared_types.UserProvisionDetails)
 	for i := range provisionDetails {
 		if provisionDetails[i].SSHKeyID != nil {
@@ -165,7 +144,6 @@ func (s *ServerStorage) ListServersByOrganizationID(orgID uuid.UUID, params type
 		}
 	}
 
-	// Aggregate resource totals (vcpu, ram, disk) per SSH key
 	type resourceAggregate struct {
 		SSHKeyID    uuid.UUID `bun:"ssh_key_id"`
 		TotalVcpu   int       `bun:"total_vcpu"`
@@ -173,7 +151,7 @@ func (s *ServerStorage) ListServersByOrganizationID(orgID uuid.UUID, params type
 		TotalDiskGB int       `bun:"total_disk_gb"`
 	}
 	var aggregates []resourceAggregate
-	err = s.getDB().NewSelect().
+	if err = s.getDB().NewSelect().
 		TableExpr("user_provision_details").
 		ColumnExpr("ssh_key_id").
 		ColumnExpr("COALESCE(SUM(vcpu_count), 0) AS total_vcpu").
@@ -182,8 +160,7 @@ func (s *ServerStorage) ListServersByOrganizationID(orgID uuid.UUID, params type
 		Where("ssh_key_id IN (?)", bun.In(sshKeyIDs)).
 		Where("organization_id = ?", orgID).
 		GroupExpr("ssh_key_id").
-		Scan(s.Ctx, &aggregates)
-	if err != nil {
+		Scan(s.Ctx, &aggregates); err != nil {
 		return nil, 0, err
 	}
 
@@ -192,33 +169,28 @@ func (s *ServerStorage) ListServersByOrganizationID(orgID uuid.UUID, params type
 		aggregateMap[agg.SSHKeyID] = agg
 	}
 
-	// Combine SSH keys with provision details and resource totals
-	servers := make([]types.ServerResponse, 0, len(sshKeys))
+	machines := make([]types.MachineResponse, 0, len(sshKeys))
 	for _, key := range sshKeys {
-		server := types.ServerResponse{
-			SSHKey: key,
-		}
+		m := types.MachineResponse{SSHKey: key}
 		if provision, ok := provisionMap[key.ID]; ok {
-			server.Provision = provision
+			m.Provision = provision
 		}
 		if agg, ok := aggregateMap[key.ID]; ok {
-			server.TotalVcpu = agg.TotalVcpu
-			server.TotalRamMB = agg.TotalRamMB
-			server.TotalDiskGB = agg.TotalDiskGB
+			m.TotalVcpu = agg.TotalVcpu
+			m.TotalRamMB = agg.TotalRamMB
+			m.TotalDiskGB = agg.TotalDiskGB
 		}
-		servers = append(servers, server)
+		machines = append(machines, m)
 	}
 
-	return servers, totalCount, nil
+	return machines, totalCount, nil
 }
 
-// GetServerByIDAndOrgID retrieves a non-deleted SSH key that belongs to the given org.
-// Returns sql.ErrNoRows if not found or wrong org.
-func (s *ServerStorage) GetServerByIDAndOrgID(serverID uuid.UUID, orgID uuid.UUID) (*shared_types.SSHKey, error) {
+func (s *ListStorage) GetMachineByIDAndOrgID(machineID uuid.UUID, orgID uuid.UUID) (*shared_types.SSHKey, error) {
 	var key shared_types.SSHKey
 	err := s.getDB().NewSelect().
 		Model(&key).
-		Where("id = ?", serverID).
+		Where("id = ?", machineID).
 		Where("organization_id = ?", orgID).
 		Where("deleted_at IS NULL").
 		Scan(s.Ctx)
@@ -228,17 +200,13 @@ func (s *ServerStorage) GetServerByIDAndOrgID(serverID uuid.UUID, orgID uuid.UUI
 	return &key, nil
 }
 
-// SetDefaultServer atomically designates serverID as the org's default server.
-// Returns the previous default's ID (nil if none existed).
-// Returns types.ErrServerNotFound if target not found/wrong org, types.ErrServerInactive if inactive.
-func (s *ServerStorage) SetDefaultServer(orgID uuid.UUID, serverID uuid.UUID) (*uuid.UUID, error) {
+func (s *ListStorage) SetDefaultMachine(orgID uuid.UUID, machineID uuid.UUID) (*uuid.UUID, error) {
 	tx, err := s.DB.BeginTx(s.Ctx, nil)
 	if err != nil {
 		return nil, err
 	}
 	defer tx.Rollback()
 
-	// Step 1: capture old default ID
 	var oldKey shared_types.SSHKey
 	var oldDefaultID *uuid.UUID
 	err = tx.NewSelect().
@@ -256,23 +224,21 @@ func (s *ServerStorage) SetDefaultServer(orgID uuid.UUID, serverID uuid.UUID) (*
 		oldDefaultID = &id
 	}
 
-	// Step 2: validate target server (must exist, be in org, and be active)
 	var target shared_types.SSHKey
 	err = tx.NewSelect().
 		Model(&target).
 		Column("id", "is_active").
-		Where("id = ?", serverID).
+		Where("id = ?", machineID).
 		Where("organization_id = ?", orgID).
 		Where("deleted_at IS NULL").
 		Scan(s.Ctx)
 	if err != nil {
-		return nil, types.ErrServerNotFound
+		return nil, types.ErrMachineNotFound
 	}
 	if !target.IsActive {
-		return nil, types.ErrServerInactive
+		return nil, types.ErrMachineInactive
 	}
 
-	// Step 3: clear existing default
 	_, err = tx.NewUpdate().
 		Model((*shared_types.SSHKey)(nil)).
 		Set("is_default = ?", false).
@@ -284,12 +250,11 @@ func (s *ServerStorage) SetDefaultServer(orgID uuid.UUID, serverID uuid.UUID) (*
 		return nil, err
 	}
 
-	// Step 4: set new default
 	_, err = tx.NewUpdate().
 		Model((*shared_types.SSHKey)(nil)).
 		Set("is_default = ?", true).
 		Set("updated_at = NOW()").
-		Where("id = ?", serverID).
+		Where("id = ?", machineID).
 		Exec(s.Ctx)
 	if err != nil {
 		return nil, err
