--- api/internal/features/trail/controller/status.go
@@ -1,57 +1,0 @@
-package controller
-
-import (
-	"github.com/go-fuego/fuego"
-	"github.com/google/uuid"
-	"github.com/nixopus/nixopus/api/internal/features/logger"
-	"github.com/nixopus/nixopus/api/internal/features/trail/types"
-	"github.com/nixopus/nixopus/api/internal/utils"
-)
-
-// GetStatus handles GET /api/v1/trail/status/{sessionId}
-//
-// This endpoint retrieves the current status of a trail provisioning request.
-// Only the owner of the provision can access its status.
-//
-// Path Parameters:
-//   - sessionId: the UUID of the provision session
-//
-// Returns:
-//   - 200 OK: status retrieved successfully
-//   - 400 Bad Request: invalid session ID format
-//   - 401 Unauthorized: authentication required
-//   - 404 Not Found: provision not found or not owned by user
-func (c *TrailController) GetStatus(f fuego.ContextNoBody) (*types.TrailStatusEnvelopeResponse, error) {
-	w, r := f.Response(), f.Request()
-	user := utils.GetUser(w, r)
-
-	if user == nil {
-		return nil, fuego.UnauthorizedError{Detail: "authentication required"}
-	}
-
-	sessionID := f.PathParam("sessionId")
-	if sessionID == "" {
-		return nil, fuego.BadRequestError{Detail: types.ErrInvalidSessionID.Error(), Err: types.ErrInvalidSessionID}
-	}
-
-	if _, err := uuid.Parse(sessionID); err != nil {
-		return nil, fuego.BadRequestError{Detail: types.ErrInvalidSessionID.Error(), Err: types.ErrInvalidSessionID}
-	}
-
-	result, err := c.service.GetStatus(user.ID.String(), sessionID)
-	if err != nil {
-		c.logger.Log(logger.Error, err.Error(), user.ID.String())
-		status := mapErrorToStatus(err)
-		return nil, fuego.HTTPError{
-			Err:    err,
-			Detail: err.Error(),
-			Status: status,
-		}
-	}
-
-	return &types.TrailStatusEnvelopeResponse{
-		Status:  "success",
-		Message: "Status retrieved successfully",
-		Data:    result,
-	}, nil
-}
