+++ api/internal/routes/trail.go
@@ -2,15 +2,15 @@ package routes
 
 import (
 	"github.com/go-fuego/fuego"
-	trail "github.com/nixopus/nixopus/api/internal/features/trail/controller"
+	machine_controller "github.com/nixopus/nixopus/api/internal/features/machine/controller"
 )
 
-func (router *Router) RegisterTrailRoutes(group *fuego.Server, controller *trail.TrailController) {
+func (router *Router) RegisterTrailRoutes(group *fuego.Server, controller *machine_controller.TrailController) {
 	fuego.Post(group, "/provision", controller.ProvisionTrail, fuego.OptionSummary("Provision trail resources"))
 	fuego.Get(group, "/status/{sessionId}", controller.GetStatus, fuego.OptionSummary("Get trail session status"))
 }
 
-func (router *Router) RegisterTrailInternalRoutes(group *fuego.Server, controller *trail.TrailController) {
+func (router *Router) RegisterTrailInternalRoutes(group *fuego.Server, controller *machine_controller.TrailController) {
 	fuego.Post(
 		group,
 		"/upgrade-resources",
